# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: FYN Social Connection API
  description: REST API for FYN Social Connection App
  version: 2.0.0
  contact:
    name: FYN Development Team
servers:
  - url: https://api.fyn.app/v2
    description: Production
  - url: http://localhost:8080/v2
    description: Development
...
paths:
...
  /matching/discover:
    get:
      tags: [Matching]
      summary: Get matched users for discovery
      security: [{ bearerAuth: [] }]
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [dating, friendship, activity, study, networking]
        - name: distance
          in: query
          schema: { type: integer, default: 50 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 50 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoverResponse'
...
  /matching/swipe:
    post:
      tags: [Matching]
      summary: Record swipe action
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [targetUserId, action]
              properties:
                targetUserId: { type: string, format: uuid }
                action: { type: string, enum: [like, pass, super_like] }
                context: { type: string, enum: [dating, friendship, activity] }
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  isMatch: { type: boolean }
                  connection:
                    $ref: '#/components/schemas/Connection'

  # ============================================
  # CONNECTIONS
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DiscoverResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/MatchedUser'
        next_cursor: { type: string }
    MatchedUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        age: { type: integer }
        avatar_url: { type: string }
        distance_km: { type: number }
        match_score: { type: number }
        matched_interests: { type: array, items: { type: string } }
        looking_for: { type: array, items: { type: string } }
        is_verified: { type: boolean }
        reputation_score: { type: number }
    Connection:
      type: object
      properties:
        id: { type: string, format: uuid }
        requester_id: { type: string, format: uuid }
        receiver_id: { type: string, format: uuid }
        connection_type: { type: string }
        status: { type: string }
        match_score: { type: number }
        created_at: { type: string, format: date-time }
...